Update des posts complété

+++ backend/controllers/posts.js
@@ -45,6 +45,43 @@ exports.getOnePost = (req, res, next) => {
 };
 
 // Mécanique de modification d'un post
+// exports.updateOnePost = (req, res, next) => {
+//   Post.findOne({
+//     where: { id: req.params.id },
+//   }).then((post) => {
+//     if (!post) {
+//       return res.status(404).json({
+//         error: new Error("Post non trouvé !"),
+//       });
+//     }
+//     if (post.userId === req.auth.userId || req.auth.isAdmin) {
+//       console.log(req.file, req.body.file, post.image, req.body.content);
+//       if (req.file && !req.body.content) {
+//         if (post.image) {
+//           const filename = post.image.split("/images/")[1];
+//           fs.unlink(`images/${filename}`, () => {});
+//         }
+//         Post.update(
+//           {
+//             image: `${req.protocol}://${req.get("host")}/images/${
+//               req.file.filename
+//             }`,
+//           },
+//           { where: { id: req.params.id } }
+//         )
+//           .then(() => {
+//             console.log("Ca fonctionne");
+//             res.status(200).json({ message: "Post mis à jour" });
+//           })
+//           .catch((error) => {
+//             console.log("ça ne fonctionne pas");
+//             res.status(401).json({ error });
+//           });
+//       }
+//     }
+//   });
+// };
+
 exports.updateOnePost = (req, res, next) => {
   Post.findOne({
     where: { id: req.params.id },
@@ -55,14 +92,14 @@ exports.updateOnePost = (req, res, next) => {
       });
     }
     if (post.userId === req.auth.userId || req.auth.isAdmin) {
-      if (req.file) {
+      console.log(req.file);
+      if (req.file && !req.body.content) {
         if (post.image) {
           const filename = post.image.split("/images/")[1];
           fs.unlink(`images/${filename}`, () => {});
         }
         Post.update(
           {
-            ...req.body,
             image: `${req.protocol}://${req.get("host")}/images/${
               req.file.filename
             }`,
@@ -72,7 +109,10 @@ exports.updateOnePost = (req, res, next) => {
           .then(() => res.status(200).json({ message: "Post mis à jour" }))
           .catch((error) => res.status(401).json({ error }));
       } else {
-        Post.update({ ...req.body }, { where: { id: req.params.id } })
+        Post.update(
+          { ...req.body, image: post.image },
+          { where: { id: req.params.id } }
+        )
           .then(() => res.status(200).json({ message: "Post mis à jour" }))
           .catch((error) => res.status(401).json({ error }));
       }
